+++ python/instrumentation/openinference-instrumentation-smolagents/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.17](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-smolagents-v0.1.16...python-openinference-instrumentation-smolagents-v0.1.17) (2025-09-11)
+
+
+### Bug Fixes
+
+* **smolagents:** Getting Multiple Traces While Streaming Code Agent ([#1872](https://github.com/Arize-ai/openinference/issues/1872)) ([420ea79](https://github.com/Arize-ai/openinference/commit/420ea79a5fd2ce572cb7bdcb38b6585107cce465))
+
 ## [0.1.16](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-smolagents-v0.1.15...python-openinference-instrumentation-smolagents-v0.1.16) (2025-08-22)